Components: Extract a reusable MultiBlocksSwitcher component
