+++ editor/edit-post/header/header-toolbar/index.js
@@ -12,12 +12,18 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import './style.scss';
-import { Inserter, BlockToolbar, TableOfContents, EditorHistoryRedo, EditorHistoryUndo } from '../../../components';
-import BlockSwitcher from '../../../components/block-switcher';
+import {
+	Inserter,
+	BlockToolbar,
+	TableOfContents,
+	EditorHistoryRedo,
+	EditorHistoryUndo,
+	MultiBlocksSwitcher,
+} from '../../../components';
 import NavigableToolbar from '../../../components/navigable-toolbar';
-import { getMultiSelectedBlockUids, isFeatureActive } from '../../../selectors';
+import { isFeatureActive } from '../../../selectors';
 
-function HeaderToolbar( { hasFixedToolbar, isMultiBlockSelection, selectedBlockUids } ) {
+function HeaderToolbar( { hasFixedToolbar } ) {
 	return (
 		<NavigableToolbar
 			className="editor-header-toolbar"
@@ -27,10 +33,7 @@ function HeaderToolbar( { hasFixedToolbar, isMultiBlockSelection, selectedBlockU
 			<EditorHistoryUndo />
 			<EditorHistoryRedo />
 			<TableOfContents />
-			{ isMultiBlockSelection && (
-				<div className="editor-header-toolbar__block-toolbar">
-					<BlockSwitcher key="switcher" uids={ selectedBlockUids } />
-				</div> ) }
+			<MultiBlocksSwitcher />
 			{ hasFixedToolbar && (
 				<div className="editor-header-toolbar__block-toolbar">
 					<BlockToolbar />
@@ -41,12 +44,7 @@ function HeaderToolbar( { hasFixedToolbar, isMultiBlockSelection, selectedBlockU
 }
 
 export default connect(
-	( state ) => {
-		const selectedBlockUids = getMultiSelectedBlockUids( state );
-		return {
-			hasFixedToolbar: isFeatureActive( state, 'fixedToolbar' ),
-			isMultiBlockSelection: selectedBlockUids.length > 1,
-			selectedBlockUids,
-		};
-	}
+	( state ) => ( {
+		hasFixedToolbar: isFeatureActive( state, 'fixedToolbar' ),
+	} )
 )( HeaderToolbar );
